+++ plugins-ee/openid-connect/kong/plugins/openid-connect/consumers.lua
@@ -116,7 +116,7 @@ local function set_anonymous(ctx, anonymous, client)
     }
   }
 
-  local consumer, err = find_consumer(consumer_token, "id", true, "id", false)
+  local consumer, err = find_consumer(consumer_token, "id", true, nil, false)
   if type(consumer) ~= "table" then
     if err then
       return unexpected(client, "anonymous consumer was not found (", err, ")")